updated reliability score to enable proof of work checks
