+++ setup1-shared/Cargo.toml
@@ -8,8 +8,14 @@ edition = "2018"
 default = []
 async_message = ["tokio"]
 twitter = ["egg-mode"]
+proof-of-work = ["phase1", "setup-utils", "snarkvm-curves"]
 
 [dependencies]
+phase1 = { path = "../phase1", features = ["parallel"], optional = true }
+setup-utils = { path = "../setup-utils", features = ["parallel"], optional = true }
+
+snarkvm-curves = { git = "https://github.com/AleoHQ/snarkVM.git", rev = "fc997c", optional = true }
+
 egg-mode = { version = "0.16", optional = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
