+++ setup1-shared/src/lib.rs
@@ -1,2 +1,4 @@
+#[cfg(feature = "proof-of-work")]
+pub mod proof_of_work;
 pub mod reliability;
 pub mod structures;
